// src/types.ts
export interface Options {
  url: string
  target: string
  config: string
  backend: string
  include: string
  exclude: string
  name: string
  emoji: 'true' | undefined
  append_type: 'true' | undefined
  append_info: 'true' | undefined
  scv: 'true' | undefined
  udp: 'true' | undefined
  list: 'true' | undefined
  sort: 'true' | undefined
  fdn: 'true' | undefined
  insert: 'true' | undefined
}